@@ -1,11 +1,14 @@
 {
   "name": "webb-universe-api",
   "version": "0.1.0",
+  "repository": {},
+  "license": "UNLICENSED",
   "scripts": {
     "lint": "gulp lint",
-    "start": "npm run serve",
-    "serve": "gulp default",
-    "pm2:start": "pm2 start ./bin/www --name 'api.webbuniverse.com' --update-env",
+    "start": "gulp default",
+    "build": "tsc",
+    "release": "standard-version",
+    "pm2:start": "npm build && pm2 start ./dist/server.js --name 'api.webbuniverse.com' --update-env",
     "pm2:restart": "pm2 restart api.webbuniverse.com",
     "pm2:stop": "pm2 stop api.webbuniverse.com",
     "pm2:delete": "pm2 delete api.webbuniverse.com",
@@ -13,46 +16,78 @@
     "pm2:logs": "pm2 logs api.webbuniverse.com"
   },
   "_moduleAliases": {
-    "@config": "config",
-    "@endpoints": "api/endpoints",
-    "@services": "api/services"
+    "@config": "src/config",
+    "@types": "src/types",
+    "@helpers": "src/helpers",
+    "@services": "src/services",
+    "@api": "src/api"
   },
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "chalk": "^3.0.0",
-    "contentful": "^7.14.0",
-    "cors": "^2.8.5",
-    "del": "^5.1.0",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "fancy-log": "^1.3.3",
-    "fs": "0.0.1-security",
+  "devDependencies": {
+    "@types/cors": "^2.8.6",
+    "@types/dotenv-safe": "^8.1.0",
+    "@types/errorhandler": "^1.5.0",
+    "@types/express": "^4.17.6",
+    "@types/express-session": "^1.17.0",
+    "@types/graphql": "^14.5.0",
+    "@types/method-override": "0.0.31",
+    "@types/morgan": "^1.9.0",
+    "@types/node": "^13.13.4",
+    "@types/uuid": "^7.0.3",
+    "@typescript-eslint/eslint-plugin": "^2.30.0",
+    "@typescript-eslint/parser": "^2.30.0",
+    "cross-env": "^7.0.2",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-import-resolver-alias": "^1.1.2",
+    "eslint-import-resolver-typescript": "^2.0.0",
+    "eslint-plugin-import": "^2.20.1",
     "gulp": "^4.0.2",
     "gulp-eslint": "^6.0.0",
     "gulp-hub": "^4.2.0",
     "gulp-nodemon": "^2.4.2",
     "gulp-rename": "^2.0.0",
+    "nodemon": "^2.0.3",
+    "standard-version": "^7.1.0",
+    "ts-node": "^8.9.1",
+    "typescript": "^3.8.3"
+  },
+  "dependencies": {
+    "@types/debug": "^4.1.5",
+    "@types/fancy-log": "^1.3.1",
+    "apollo-server-express": "^2.12.0",
+    "body-parser": "^1.19.0",
+    "chalk": "^3.0.0",
+    "contentful": "^7.14.0",
+    "core-util-is": "^1.0.2",
+    "cors": "^2.8.5",
+    "dataloader": "^2.0.0",
+    "debug": "^4.1.1",
+    "del": "^5.1.0",
+    "dotenv-safe": "^8.2.0",
+    "ejs": "^3.1.2",
+    "errorhandler": "^1.5.1",
+    "express": "^4.17.1",
+    "express-session": "^1.17.1",
+    "fancy-log": "^1.3.3",
+    "fs": "0.0.1-security",
+    "graphql": "^14.2.1",
     "http": "0.0.0",
+    "i": "^0.3.6",
     "lodash": "^4.17.15",
     "lodash.get": "^4.4.2",
+    "method-override": "^3.0.0",
     "module-alias": "^2.2.2",
-    "moment": "^2.24.0",
-    "mongoose": "^5.9.2",
     "morgan": "^1.9.1",
-    "node-sass": "^4.13.1",
-    "nodemon": "^2.0.2",
     "path": "^0.12.7",
     "pm2": "^4.2.3",
+    "reflect-metadata": "^0.1.13",
     "request": "^2.88.2",
     "rfs": "^9.0.2",
-    "rotating-file-stream": "^2.0.2",
+    "rotating-file-stream": "^2.1.0",
     "stream": "0.0.2",
+    "type-graphql": "^0.17.6",
+    "typeorm": "^0.2.24",
+    "uuid": "^8.0.0",
     "winston": "^3.2.1"
-  },
-  "devDependencies": {
-    "eslint": "^6.1.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.20.1"
   }
 }
