@@ -8,12 +8,12 @@ const esLintOpts = {
 
 /** Lints server JS */
 const lint = () => {
-  logger.message('Linting Server', 'blue.bold');
+  logger.info('Linting Server', 'blue.bold');
 
-  return gulp.src([`api/**/*.js`])
+  return gulp.src([`src/**/*.ts`])
     .pipe(eslint(esLintOpts))
     .pipe(eslint.format())
     .pipe(eslint.failAfterError());
-}
+};
 
 module.exports = { lint };
